Fixed a bug in SharedUsageLimitProxyMemoryAllocator

+++ util/SharedUsageLimitProxyMemoryAllocator.h
@@ -69,10 +69,10 @@ public:
 
 private:
    void ResetDaemonCounter(); // Note: this assumes the SharedMemory is already locked for read/write!
-   status_t ChangeDaemonCounter(int32 byteDelta);  // Note: this assumes the SharedMemory is not locked yet
-   status_t ChangeDaemonCounterAux(int32 byteDelta);  // Note: this assumes the SharedMemory is not locked yet
+   status_t ChangeDaemonCounter(   ssize_t byteDelta);  // Note: this assumes the SharedMemory is not locked yet
+   status_t ChangeDaemonCounterAux(ssize_t byteDelta);  // Note: this assumes the SharedMemory is not locked yet
    MUSCLE_NODISCARD size_t CalculateTotalAllocationSum() const;
-   MUSCLE_NODISCARD uint32 GetNumSlots() const {return _shared.GetAreaSize()/sizeof(size_t);}
+   MUSCLE_NODISCARD size_t GetNumSlots() const {return _shared.GetAreaSize()/sizeof(size_t);}
    MUSCLE_NODISCARD size_t * GetArrayPointer() {return reinterpret_cast<size_t *>(_shared());}
    MUSCLE_NODISCARD const size_t * GetArrayPointer() const {return reinterpret_cast<const size_t *>(_shared());}
